user_threads: Add click to scroll to next comment from `@user`
# Motivation

When reading the comments of a user, sometimes the comment is part of a
large thread of comments. Making it hard to find the next thread the
user participated in. Implement similar functionality to the one already
present to navigate unread comments

# Changes

- Add event handler to elements with `showing-user-comment` class that
  scrolls to the next one when pressed.
- Add cursor: pointer to the elements with `showing-user-comment` class.
- Add role='link' to the elements with `showing-user-comment` class.

# Notes

The event handler code is literally a copy & paste of the previous event
handler. We could try to factor the code, parameterizing by the css selector to
find the related elements. I don't think it is worth it in this case. The code
is small & co-located. So its easy to apply changes to both handlers.